+++ src/ws/slts/sentence.py
