+++ src/ws/stsl/text.py
